[Android System Bars] Rewritted the code;
The previous solution was not secure and it was not the official solution given by android. There were reports of failures on cell phones like HTC. In addition, it only worked correctly in portrait mode, because NavigationBar would often go to the left of the screen or to the right of the screen and we had no way to deal with it.